[Enhancement] #75 Clean structure includeing move some functionality from PanelProvider ro FileService
