+++ src/services/panelProvider.js
@@ -1,30 +1,25 @@
+const { removeCommentStructure } = require("../helpers/codeHelper");
+const { uri, load } = require("../helpers/fileSystemHelper");
+const { callGitDiscard } = require("./gitService");
+const FileService = require("./fileService");
 const vscode = require("vscode");
-const path = require("path");
-const fs = require("fs");
-
 const {
   openCodeFile,
   copyCodeBlock,
   sendToTerminal,
-  replaceActiveFile,
 } = require("../commands/commands");
 const {
   undoCodeBlock,
   diffCodeBlock,
   buildProjectStructure,
 } = require("../helpers/extensionHelper");
-const {
-  uri,
-  load,
-} = require("../helpers/fileSystemHelper");
-const { callGitDiscard } = require("./gitService");
-const { removeCommentStructure } = require("../helpers/codeHelper");
 
 class PanelProvider {
   constructor(extensionUri) {
     this._extensionUri = extensionUri;
     this._backupFilePath = null;
     this._view = null;
+    this._fileService = new FileService();
   }
 
   async resolveWebviewView(webviewView) {
@@ -46,10 +41,10 @@ class PanelProvider {
         break;
       case "replaceActiveFile":
         await openCodeFile(message.code);
-        await this._replaceActiveFile(message.code);
+        await this._fileService.replaceActiveFile(message.code);
         break;
       case "replaceCodeFileSilently":
-        await this._replaceCodeFileSilently(message.code);
+        await this._fileService.replaceCodeFileSilently(message.code);
         break;
       case "copyCodeBlock":
         await copyCodeBlock(removeCommentStructure(message.code));
@@ -69,89 +64,6 @@ class PanelProvider {
       case "sendToTerminal":
         await sendToTerminal(message.code);
         break;
-    }
-  }
-
-  async _cloneAndModifyActiveFile(code, modifyFunction) {
-    const editor = vscode.window.activeTextEditor;
-    if (!editor) {
-      vscode.window.showErrorMessage("No active editor.");
-      return;
-    }
-
-    const filePath = editor.document.uri.fsPath;
-    const workspaceFolder = vscode.workspace.getWorkspaceFolder(
-      editor.document.uri
-    );
-
-    if (!workspaceFolder) {
-      vscode.window.showErrorMessage("File is not part of a workspace.");
-      return;
-    }
-
-    const relativePath = path.relative(workspaceFolder.uri.fsPath, filePath);
-    const backupDir = path.join(workspaceFolder.uri.fsPath, ".n8x");
-    const backupPath = path.join(backupDir, relativePath);
-
-    if (!fs.existsSync(backupDir)) {
-      fs.mkdirSync(backupDir, { recursive: true });
-    }
-
-    const backupFileDir = path.dirname(backupPath);
-    if (!fs.existsSync(backupFileDir)) {
-      fs.mkdirSync(backupFileDir, { recursive: true });
-    }
-
-    if (!fs.existsSync(backupPath)) {
-      try {
-        fs.writeFileSync(backupPath, editor.document.getText());
-        console.log(`Backup file created: ${backupPath}`);
-      } catch (err) {
-        vscode.window.showErrorMessage(
-          `Failed to create backup: ${err.message}`
-        );
-        return;
-      }
-    }
-
-    await modifyFunction(code);
-  }
-
-  async _replaceActiveFile(code) {
-    await this._cloneAndModifyActiveFile(code, async (modifiedCode) => {
-      await replaceActiveFile(modifiedCode);
-    });
-  }
-
-  async _replaceCodeFileSilently(code) {
-    const filePathMatch = code.match(/^.*?\[\[(.*?)\]\]/);
-    if (!filePathMatch || !filePathMatch[1]) {
-      return vscode.window.showErrorMessage(
-        "File path not found in the code block."
-      );
-    }
-    let filePath = filePathMatch[1].trim();
-
-    // Check if the file path is absolute or relative
-    if (!path.isAbsolute(filePath)) {
-      // If the file path is relative make it absolute relative to workspace.
-      const workspaceFolders = vscode.workspace.workspaceFolders;
-      if (!workspaceFolders || workspaceFolders.length === 0) {
-        return vscode.window.showErrorMessage(
-          "No workspace folder open to resolve relative path."
-        );
-      }
-      filePath = path.join(workspaceFolders[0].uri.fsPath, filePath);
-    }
-
-    try {
-      const modifiedCode = removeCommentStructure(code);
-      fs.writeFileSync(filePath, modifiedCode);
-      console.log(`File replaced silently: ${filePath}`);
-    } catch (err) {
-      vscode.window.showErrorMessage(
-        `Failed to replace file content silently: ${err.message}`
-      );
     }
   }
